refactor: Split influxrpc end to end tests into smaller modules
